+++ handlers/federation/sql_driver.go
@@ -20,27 +20,7 @@ func (drv *SQLDriver) LookupRecord(name, domain string) (*Record, error) {
 	return &result, nil
 }
 
-// LookupReverseRecord implements `ReverseDriver` by performing
-// `drv.LookupReverseRecordQuery` against `drv.DB` using the provided parameter
-func (drv *SQLDriver) LookupReverseRecord(
-	accountid string,
-) (*ReverseRecord, error) {
-	drv.initDB()
-	var result ReverseRecord
-
-	err := drv.db.GetRaw(&result, drv.LookupReverseRecordQuery, accountid)
-
-	if drv.db.NoRows(err) {
-		return nil, nil
-	} else if err != nil {
-		return nil, errors.Wrap(err, "db get")
-	}
-
-	return &result, nil
-}
-
 var _ Driver = &SQLDriver{}
-var _ ReverseDriver = &SQLDriver{}
 
 func (drv *SQLDriver) initDB() {
 	drv.init.Do(func() {